+++ src/downloading/dumpFurni.ts
@@ -6,29 +6,34 @@ import { extractSwf } from "./extractSwf";
 
 export async function dumpFurni(
   dcrUrl: string,
-  revision: string,
+  revision: string | undefined,
   name: string,
   folder: string
 ) {
-  const dir = path.join(folder, revision, name);
+  const furniDirectory =
+    revision != null
+      ? path.join(folder, revision, name)
+      : path.join(folder, name);
+  const revisionDirectory =
+    revision != null ? path.join(folder, revision) : path.join(folder);
 
   try {
-    await fs.stat(dir);
+    await fs.stat(furniDirectory);
     return;
   } catch (e) {}
 
   const data = await fetchFurni(dcrUrl, revision, name);
 
-  await fs.mkdir(path.join(folder, revision, data.name), { recursive: true });
+  await fs.mkdir(furniDirectory, { recursive: true });
 
   await fs.writeFile(
-    path.join(folder, data.revision, `${data.name}.swf`),
+    path.join(revisionDirectory, `${data.name}.swf`),
     data.blob,
     "binary"
   );
 
   await extractSwf(
-    path.join(folder, revision, data.name),
-    path.join(folder, data.revision, `${data.name}.swf`)
+    furniDirectory,
+    path.join(revisionDirectory, `${data.name}.swf`)
   );
 }